made position f32

+++ src/context.rs
@@ -1,6 +1,6 @@
 #[derive(Clone)]
 pub struct Context {
-    pub viewing_distance: i32,
+    pub viewing_distance: f32,
     pub width: i32,
     pub height: i32,
 }
@@ -8,7 +8,7 @@ pub struct Context {
 impl Default for Context {
     fn default() -> Context {
         Context {
-            viewing_distance: 25,
+            viewing_distance: 25.0,
             width: 640,
             height: 480,
         }
